nfc: convert st25r3911b library and samples to devicetree
These samples are using Kconfig to configure a SPI device, instead of
devicetree. Using Kconfig in this way is something Zephyr has started
moving away from, and is now actively discouraging.

To ensure that all users of the library must update, require at least
one node in the devicetree with the appropriate
compatible ("st,st25r3911b") to be defined in order for
CONFIG_ST25R3911B_LIB to be enabled. That way, the Kconfig step of the
build system will scream and die if a user attempts to build the
library with an out-of-date devicetree.

This library is not a true device driver, since there is no NFC API.
Therefore, do not allocate a struct device in st25r3911b_spi.c even
though we are using devicetree.

Update the tag_reader sample overlay to provide the necessary SPI
device nodes.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>